Endre typen på TidslinjePeriode::lengde til Long i stedet for Int

+++ tidslinje/src/main/kotlin/no/nav/familie/tidslinje/utvidelser/PeriodeUtvidelser.kt
@@ -11,7 +11,7 @@ import no.nav.familie.tidslinje.TidslinjePeriode
  */
 fun <T, R, RESULTAT> TidslinjePeriode<T>.biFunksjon(
     operand: TidslinjePeriode<R>,
-    lengde: Int,
+    lengde: Long,
     erUendelig: Boolean,
     operator: (elem1: PeriodeVerdi<T>, elem2: PeriodeVerdi<R>) -> PeriodeVerdi<RESULTAT>,
 ): TidslinjePeriode<RESULTAT> = TidslinjePeriode(operator(this.periodeVerdi, operand.periodeVerdi), lengde, erUendelig)
@@ -19,7 +19,7 @@ fun <T, R, RESULTAT> TidslinjePeriode<T>.biFunksjon(
 fun <A, B, C, RESULTAT> TidslinjePeriode<A>.biFunksjon(
     operand1: TidslinjePeriode<B>,
     operand2: TidslinjePeriode<C>,
-    lengde: Int,
+    lengde: Long,
     erUendelig: Boolean,
     operator: (PeriodeVerdi<A>, PeriodeVerdi<B>, PeriodeVerdi<C>) -> PeriodeVerdi<RESULTAT>,
 ): TidslinjePeriode<RESULTAT> =
